# docs/swagger.yaml
definitions:
...
  github_com_Orken1119_HelpNet_internal_models.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      metadata:
        $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Properties'
    type: object
  github_com_Orken1119_HelpNet_internal_models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorDetail'
        type: array
    type: object
  github_com_Orken1119_HelpNet_internal_models.Event:
    properties:
      address:
        type: string
      end_data:
        type: string
      event_name:
        type: string
      finished:
        type: boolean
      id:
        type: integer
      information:
        type: string
      members_count:
        type: integer
      members_info:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.VolunteerMainInfo'
        type: array
      neccessary_people_count:
        type: integer
      organization_id:
        type: integer
      poster_url:
        type: string
      preview_url:
        type: string
      skill_direction:
        type: string
      start_data:
        type: string
    type: object
  github_com_Orken1119_HelpNet_internal_models.EventForCreating:
    properties:
      address:
        type: string
      end_data:
        type: string
      event_name:
        type: string
      information:
        type: string
      neccessary_people_count:
        type: integer
      organization_id:
        type: integer
      poster_url:
        type: string
      preview_url:
        type: string
      skill_direction:
        type: string
      start_data:
        type: string
    type: object
  github_com_Orken1119_HelpNet_internal_models.EventForEditing:
    properties:
      address:
        type: string
      end_data:
        type: string
      event_name:
        type: string
      information:
        type: string
      neccessary_people_count:
        type: integer
      poster_url:
        type: string
      preview_url:
        type: string
      skill_direction:
        type: string
      start_data:
        type: string
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.OrganizationProfile:
    properties:
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      events:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      finished_ivents:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      id:
        type: integer
      information:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      poster_url:
        type: string
      years_of_experience:
        type: integer
    type: object
  github_com_Orken1119_HelpNet_internal_models.OrganizationProfileEditing:
    properties:
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      information:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      poster_url:
        type: string
      years_of_experience:
        type: integer
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.Properties:
    properties:
      additionalProp1:
        type: string
      additionalProp2:
        type: string
      additionalProp3:
        type: string
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.SuccessResponse:
    properties:
      metadata:
        $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Properties'
      result: {}
    type: object
  github_com_Orken1119_HelpNet_internal_models.VolunteerMainInfo:
    properties:
      age:
        type: integer
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      grade:
        type: integer
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      photo_url:
        type: string
      skills:
        type: string
    type: object
...
host: localhost:2000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HelpNet API
  version: "1.0"
paths:
...
  /events/create-event:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.EventForCreating'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
  /events/delete-event/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  type: string
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
  /events/finish/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: organization finished event successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
  /events/get-event-by-direction/{direction}:
    get:
      consumes:
      - application/json
      parameters:
      - description: direction
        in: path
        name: directioin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
  /events/get-event-by-id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
...
  /events/participate-event/{userID}/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: userID
        in: path
        name: userID
        required: true
        type: integer
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully joined the event
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - event
  /events/update-event/{id}:
    put:
      consumes:
      - application/json
      description: Update an event's details by its ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Event details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.EventForEditing'
      produces:
      - application/json
      responses:
        "200":
          description: Event created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing event
      tags:
      - event
  /organizations/delete-organizations/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  type: string
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - organization
  /organizations/edit-organization-profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.OrganizationProfileEditing'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.OrganizationProfileEditing'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - organization
  /organizations/get-all-organizations-profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.OrganizationProfile'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - organization
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"